[ADF-3671] Rename Widget Controllers
