+++ e2e/pages/adf/process_services/widgets/containerWidget.ts
@@ -18,7 +18,7 @@
 import FormFields = require('../formFields');
 import { by } from 'protractor';
 
-export class Container {
+export class ContainerWidget {
 
     formFields = new FormFields();
