+++ e2e/pages/adf/process_services/widgets/displayTextWidget.ts
@@ -16,22 +16,24 @@
  */
 
 import FormFields = require('../formFields');
-import { by, element } from 'protractor';
-import Util = require('../../../../util/util');
+import { by } from 'protractor';
 
-export class Hyperlink {
+export class DisplayTextWidget {
 
     formFields = new FormFields();
-
-    fieldLocator = by.css('div[class="adf-hyperlink-widget "] a');
-
-    getFieldText(fieldId) {
-        return this.formFields.getFieldText(fieldId, this.fieldLocator);
-    }
+    labelLocator = by.css('div[class*="adf-display-text-widget"]');
+    inputLocator = by.css('input');
 
     getFieldLabel(fieldId) {
-        let label = element(by.css(`adf-form-field div[id="field-${fieldId}-container"] label`));
-        Util.waitUntilElementIsVisible(label);
-        return label.getText();
+        return this.formFields.getFieldLabel(fieldId, this.labelLocator);
     }
+
+    getFieldValue(fieldId) {
+        return this.formFields.getFieldValue(fieldId, this.inputLocator);
+    }
+
+    getFieldText(fieldId) {
+        return this.formFields.getFieldText(fieldId, this.labelLocator);
+    }
+
 }
