+++ e2e/pages/adf/process_services/widgets/displayValueWidget.ts
@@ -18,37 +18,18 @@
 import FormFields = require('../formFields');
 import { by } from 'protractor';
 
-export class Text {
+export class DisplayValueWidget {
 
     formFields = new FormFields();
-
-    labelLocator = by.css("label[class*='adf-label']");
+    labelLocator = by.css('span[class*="unknown-text"]');
+    inputLocator = by.css('input');
 
     getFieldLabel(fieldId) {
         return this.formFields.getFieldLabel(fieldId, this.labelLocator);
     }
 
-    getFieldPlaceHolder(fieldId) {
-        return this.formFields.getFieldPlaceHolder(fieldId);
-    }
-
-    setValue(fieldId, value) {
-        return this.formFields.setFieldValue(by.id, fieldId, value);
-    }
-
     getFieldValue(fieldId) {
-        return this.formFields.getFieldValue(fieldId);
+        return this.formFields.getFieldValue(fieldId, this.inputLocator);
     }
 
-    getErrorMessage(fieldId) {
-        return this.formFields.getFieldErrorMessage(fieldId);
-    }
-
-    isWidgetVisible(fieldId) {
-        return this.formFields.checkWidgetIsVisible(fieldId);
-    }
-
-    isWidgetNotVisible(fieldId) {
-        return this.formFields.checkWidgetIsHidden(fieldId);
-    }
 }
