+++ e2e/pages/adf/process_services/widgets/hyperlinkWidget.ts
@@ -16,16 +16,22 @@
  */
 
 import FormFields = require('../formFields');
-import { by } from 'protractor';
+import { by, element } from 'protractor';
+import Util = require('../../../../util/util');
 
-export class Header {
+export class HyperlinkWidget {
 
     formFields = new FormFields();
 
-    labelLocator = by.css('span[id="container-header-label"]');
+    fieldLocator = by.css('div[class="adf-hyperlink-widget "] a');
 
-    getFieldLabel(fieldId) {
-        return this.formFields.getFieldLabel(fieldId, this.labelLocator);
+    getFieldText(fieldId) {
+        return this.formFields.getFieldText(fieldId, this.fieldLocator);
     }
 
+    getFieldLabel(fieldId) {
+        let label = element(by.css(`adf-form-field div[id="field-${fieldId}-container"] label`));
+        Util.waitUntilElementIsVisible(label);
+        return label.getText();
+    }
 }
